chapter_003/src/main/java/ru/job4j/bank/Account.java
package ru.job4j.bank;

import java.util.Objects;

public class Account {
    private double value;
    private String requisites;

    public Account(double value, String requisites) {
        this.value = value;
        this.requisites = requisites;
    }

    public boolean transfer(Account destination, double amount) {
        boolean result = false;
        if (amount > 0 && this.value >= amount && destination != null) {
            result = true;
            this.value -= amount;
            destination.value += amount;
        }
        return result;
    }

    public double getValue() {
        return value;
    }

    public String getRequisites() {
        return requisites;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        Account account = (Account) o;
        return requisites.equals(account.requisites);
    }

    @Override
    public int hashCode() {
        return Objects.hash(requisites);
    }
}
